move over BackstageSignInResult, BackstageIdentityResponse, BackstageUserIdentity
