+++ plugins/auth-backend/src/lib/oauth/types.ts
@@ -16,11 +16,8 @@
 
 import express from 'express';
 import { Profile as PassportProfile } from 'passport';
-import {
-  RedirectInfo,
-  BackstageSignInResult,
-  ProfileInfo,
-} from '../../providers/types';
+import { BackstageSignInResult } from '@backstage/plugin-auth-node';
+import { RedirectInfo, ProfileInfo } from '../../providers/types';
 
 /**
  * Common options for passport.js-based OAuth providers
