+++ plugins/auth-backend/src/providers/index.ts
@@ -48,13 +48,6 @@ export type {
 
 // These types are needed for a postMessage from the login pop-up
 // to the frontend
-export type {
-  AuthResponse,
-  BackstageIdentity,
-  BackstageUserIdentity,
-  BackstageIdentityResponse,
-  BackstageSignInResult,
-  ProfileInfo,
-} from './types';
+export type { AuthResponse, BackstageIdentity, ProfileInfo } from './types';
 
 export { prepareBackstageIdentityResponse } from './prepareBackstageIdentityResponse';
