+++ plugins/auth-node/api-report.md
@@ -3,6 +3,29 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
+import { Entity } from '@backstage/catalog-model';
+
+// @public
+export interface BackstageIdentityResponse extends BackstageSignInResult {
+  identity: BackstageUserIdentity;
+}
+
+// @public
+export interface BackstageSignInResult {
+  // @deprecated
+  entity?: Entity;
+  // @deprecated
+  id: string;
+  token: string;
+}
+
+// @public
+export type BackstageUserIdentity = {
+  type: 'user';
+  userEntityRef: string;
+  ownershipEntityRefs: string[];
+};
+
 // @public
 export function getBearerTokenFromAuthorizationHeader(
   authorizationHeader: unknown,
