+++ plugins/auth-node/package.json
@@ -20,6 +20,7 @@
   },
   "dependencies": {
     "@backstage/backend-common": "^0.10.7-next.0",
+    "@backstage/catalog-model": "^0.9.10",
     "@backstage/config": "^0.1.13",
     "winston": "^3.2.1"
   },
