+++ plugins/permission-node/package.json
@@ -32,7 +32,7 @@
     "@backstage/backend-common": "^0.10.7-next.0",
     "@backstage/config": "^0.1.13",
     "@backstage/errors": "^0.2.0",
-    "@backstage/plugin-auth-backend": "^0.10.0-next.0",
+    "@backstage/plugin-auth-node": "^0.0.0",
     "@backstage/plugin-permission-common": "^0.4.0",
     "@types/express": "^4.17.6",
     "express": "^4.17.1",
